CHANGE xdebug-disabled.sh improvmentof shell command

+++ phpdocker/php-fpm/xdebug-disabled.sh
@@ -1,2 +1,6 @@
 #!/bin/bash
-mv /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini.disabled
+if [ -f /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini ]; then
+    mv /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini.disabled
+elif [ -f /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini.disabled ]; then
+    mv /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini.disabled /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini
+fi
